pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent2"
version = "0.1.0"
requires-python = ">=3.10"
description = "Minimal, stable baseline for LiveKit agent with provider checks and TTS dry-run"

dependencies = [
  "python-dotenv>=1.0.1",
  "httpx>=0.27.2",
  "aiohttp>=3.9",
  "livekit-agents[openai]~=1.2",
  "livekit-plugins-groq~=1.2",
  "livekit-plugins-deepgram>=1.2.6",
  "livekit-plugins-elevenlabs>=1.2.6",
]

[dependency-groups]
dev = [
  "ruff",
]

[tool.ruff]
line-length = 100
